Perform a second kill at the end of runtime only for the case of having interval output

+++ scripts/pktgen_bench/xmit_multiqueue.sh
@@ -137,10 +137,13 @@ function setup_timeout() {
   if (( TIMEOUT  > 0 )); then
     sleep "$TIMEOUT"
 
-    # the first kill can be futile if at the same time a new
-    # sleep interval is brought up.
     pkill -INT --parent "$$"
-    pkill -INT --parent "$$"
+
+    if (( INTERVAL > 0 )); then
+      # the first kill can be futile if at the same time a new
+      # sleep interval is brought up.
+      pkill -INT --parent "$$"
+    fi
   fi
 }
